create-firebase

+++ assets/Firebase Template/functions/refUpdate.mjs
@@ -1,14 +1,21 @@
 import { execSync } from 'child_process';
 import fs from 'fs';
 
-fs.rmSync('./src/ref', { recursive: true });
+import alias from '../functions/alias.mjs';
+
+if (fs.existsSync('./src/ref')) {
+  fs.rmSync('./src/ref', { recursive: true });
+}
+
 fs.mkdirSync('./src/ref');
 
 // TSConfig `paths`
 
 const tsconfigJson = JSON.parse(fs.readFileSync('./tsconfig.json', 'utf-8'));
 
-tsconfigJson.compilerOptions.paths = {};
+tsconfigJson.compilerOptions.paths = {
+  ...alias,
+};
 
 fs.writeFileSync('./tsconfig.json', JSON.stringify(tsconfigJson, null, 2));
